requirements-dev.txt
astroid==2.3.1
colorama==0.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.6
django-filters==0.2.1
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.2.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.2
python-decouple==3.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
typed-ast==1.4.0
wrapt==1.11.2
